requirements.txt
Django==3.1.5
djangorestframework==3.12.2
python-dotenv==0.15.0